B.1.1.0.0  Changes to be committed: 	modified:   README.md 	modified:   RTP_EXTRACT/App.config 	modified:   RTP_EXTRACT/Program.cs 	modified:   RTP_EXTRACT/RTP_EXTRACT.csproj 	modified:   RTP_EXTRACT/RTPextract.cs 	modified:   RTP_REPACK/App.config 	modified:   RTP_REPACK/FinalFile.cs 	modified:   RTP_REPACK/Program.cs 	modified:   RTP_REPACK/RTP_REPACK.csproj 	modified:   RTP_REPACK/RTPrepack.cs 	modified:   RTP_REPACK/TreeNode.cs 	modified:   RTP_REPACK/TreeStructure.cs

+++ RTP_REPACK/FinalFile.cs
@@ -78,10 +78,10 @@ namespace RTP_REPACK
 
             for (int i = 0; i < Block1Array.Length; i++)
             {
-                byte[] x = BitConverter.GetBytes(Block1Array[i].X);
-                byte[] y = BitConverter.GetBytes(Block1Array[i].Y);
-                byte[] z = BitConverter.GetBytes(Block1Array[i].Z);
-                byte[] w = BitConverter.GetBytes(1.0f);
+                byte[] x = BitConverter.GetBytes((float)Block1Array[i].X);
+                byte[] y = BitConverter.GetBytes((float)Block1Array[i].Y);
+                byte[] z = BitConverter.GetBytes((float)Block1Array[i].Z);
+                byte[] w = BitConverter.GetBytes((float)1.0f);
 
                 byte[] index = BitConverter.GetBytes((ushort)Block1Array[i].ConnectionTableIndex);
                 byte[] count = BitConverter.GetBytes((ushort)Block1Array[i].ConnectionCount);
